style: ruff + black clean across PR files (resolves format/code-quality CI)
Apply ruff lint auto-fixes (redundant int()/dict() calls, import order,
datetime UTC modernisation) and black formatting across the PR's files. Also
reformats four base #262 files (symbolic_constraint.py, neurosymbolic_ablation.py,
test_binary_conformal.py, test_fusion_symbolic_cotraining.py) that were
black-dirty before this session under the pinned black (26.x) version -- a
mechanical, semantics-preserving reformat so the PR's format + Code Quality
checks pass.

No behaviour change: ruff check (src/ tests/) and black --check are clean;
mypy clean; 39 affected tests still pass.

+++ benchmarks/domain_encoder_ablation.py
@@ -53,10 +53,10 @@ import torch
 _HERE = Path(__file__).resolve().parent
 sys.path.insert(0, str(_HERE.parent / "src"))
 
-from omni_mercury_engine.datasets.adbench import ADBenchLoader  # noqa: E402
-from omni_mercury_engine.datasets.base import DatasetConfig  # noqa: E402
-from omni_mercury_engine.engine import OmniMercuryEngine  # noqa: E402
-from omni_mercury_engine.ml.mercury_ml import roc_auc_score  # noqa: E402
+from omni_mercury_engine.datasets.adbench import ADBenchLoader
+from omni_mercury_engine.datasets.base import DatasetConfig
+from omni_mercury_engine.engine import OmniMercuryEngine
+from omni_mercury_engine.ml.mercury_ml import roc_auc_score
 
 DEFAULT_DATASETS = ["cardio", "Pima", "thyroid"]
 DEFAULT_SEEDS = [0, 1, 2]
@@ -141,7 +141,7 @@ def _stratified_split(
     for cls in np.unique(y):
         idx = np.where(y == cls)[0]
         rng.shuffle(idx)
-        cut = max(1, int(round(len(idx) * train_frac)))
+        cut = max(1, round(len(idx) * train_frac))
         tr.extend(idx[:cut].tolist())
         te.extend(idx[cut:].tolist())
     return np.array(tr), np.array(te)
